REFACTOR :: GoodsList (#40)
- 각 page 겹치는 부분 GoodsList component로 추출
- GoodsList에 sortKey 부분 json으로 뺌
- 각 페이지 데이터에 sortKey 추가
- GoodsList sortKey 선택된 것 색이랑 이런거 변경
- thrifty_shopping -> thriftyShopping (카멜케이스)
