+++ src/components/GoodsList/GoodsList.scss
@@ -7,7 +7,7 @@ section {
 
     .section-header-title {
       .section-title {
-        color: rgb(51, 51, 51);
+        color: $dark-gray;
         font-size: 28px;
         letter-spacing: -0.26px;
         font-weight: 500;
@@ -43,6 +43,12 @@ section {
               a {
                 color: $light-gray;
                 cursor: pointer;
+              }
+
+              .checked {
+                color: $dark-gray;
+                letter-spacing: -0.3px;
+                font-weight: 500;
               }
             }
