+++ src/components/GoodsList/sortKeyData.json
@@ -1,0 +1,1 @@
+["추천순", "신상품순", "판매량순", "혜택순", "낮은 가격순", "높은 가격순"]
