--- src/pages/Thrifty_shopping/Thrifty_shopping.scss
@@ -1,60 +1,0 @@
-@import "../../styles/common.scss";
-
-section {
-  .section-header {
-    width: 100%;
-    padding: 50px 0 20px;
-
-    .section-header-title {
-      .section-title {
-        color: rgb(51, 51, 51);
-        font-size: 28px;
-        letter-spacing: -0.26px;
-        font-weight: 500;
-      }
-    }
-  }
-
-  .big-product-contents {
-    align-items: center;
-    width: 100%;
-
-    .products_bar {
-      display: flex;
-      justify-content: space-between;
-      padding: 18px 0 16px;
-
-      .total_count {
-        span {
-          color: $dark-gray;
-          font-size: 12px;
-          font-weight: 500;
-        }
-      }
-
-      .products_button_total {
-        .total_button {
-          .button_list {
-            li {
-              align-items: center;
-              margin-left: 8px;
-              font-size: 12px;
-
-              a {
-                color: $light-gray;
-                cursor: pointer;
-              }
-            }
-
-            .bar_before::before {
-              content: "|";
-              height: 10px;
-              margin-right: 8px;
-              color: $light-gray;
-            }
-          }
-        }
-      }
-    }
-  }
-}
